requirements.txt

Werkzeug==3.1.3
gunicorn==23.0.0
python-dotenv==1.1.1
Flask==3.1.2
PyJWT==2.10.1
bleach==6.2.0
